+++ internal/storagenode/admin_server.go
@@ -75,3 +75,7 @@ func (as *adminServer) Sync(ctx context.Context, req *snpb.SyncRequest) (*snpb.S
 func (as *adminServer) GetPrevCommitInfo(context.Context, *snpb.GetPrevCommitInfoRequest) (*snpb.GetPrevCommitInfoResponse, error) {
 	panic("not implemented")
 }
+
+func (as *adminServer) Trim(context.Context, *snpb.TrimRequest) (*snpb.TrimResponse, error) {
+	panic("not implemented")
+}