+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `--server-max-msg-size` option to storage node. (#VARLOG-688/#601)
 - Added `--logdir` and `--logtostderr` options to varlogadm. (#VARLOG-692/#604)
 - Added Trim RPC to package `proto/vmspb`. (#VARLOG-706/#608)
+- Added Trim RPC to `proto/snpb/management.proto`. (#VARLOG-708/#609)
+- Added `Trim` method to `pkg/snc.(StorageNodeManagementClient)`. (#VARLOG-708/#609)
+- Added `Trim` method to `internal/storagenode.(adminServer)`. (#VARLOG-708/#609)
 
 ### Changed
 - Changed default storage settings for good performance in usual cases. (#VARLOG-668/#584)